+++ nat/src/stateful/mod.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Open Network Fabric Authors
 
-pub(crate) mod allocator;
+pub(crate) mod allocation;
 mod allocator_writer;
 pub mod apalloc;
 pub(crate) mod flows;
@@ -15,7 +15,7 @@ pub use allocator_writer::NatAllocatorWriter;
 pub use allocator_writer::StatefulNatConfig;
 
 use super::NatTranslationData;
-use crate::stateful::allocator::{AllocationResult, AllocatorError};
+use crate::stateful::allocation::{AllocationResult, AllocatorError};
 use crate::stateful::allocator_writer::NatAllocatorReader;
 use crate::stateful::apalloc::{AllocatedIpPort, NatIpWithBitmap};
 use crate::stateful::natip::NatIp;
